Update README.md
SERVER INFO:

This is a server made for fun, roleplaying and community purposes, it is not a military-related server and people can do whatever they want except for breaking the rules of the server. This server was designed to have a friendly environment with people roleplaying as STAR WARS characters. 

STAR WARS ERA INFO:

Early Imperial Era
The era is 19 BBY - 14 BBY. 
The Empire has started to control the galaxy from the interior of the galaxy to the outermost parts of the galaxy. Due to the purge of the Jedi, the Empire has successfully taken over the galaxy and is starting to getting rid of the Republic insurgents that are plotting against the Emperor.